+++ hoomd/md/BondTablePotentialGPU.cu
@@ -35,7 +35,7 @@
 */
 __global__ void gpu_compute_bondtable_forces_kernel(Scalar4* d_force,
                                      Scalar* d_virial,
-                                     const unsigned int virial_pitch,
+                                     const size_t virial_pitch,
                                      const unsigned int N,
                                      const Scalar4 *d_pos,
                                      const BoxDim box,
